Feature: Introduce tag heirarchies.
This will probably be the method used to create instances of things - for instance, instances of objects, furniture, or people.
For example, to define a new person "Tesla", you would create a tag "Tesla", and define its parents "Serbian|Human|Physicist", define "Serbian" parents as "Nationality", "Human" parents as "Species", "Physicist"'s parents as "Scientist", "Nationality"'s parents as nothing, "Species"'s parents as nothing, and "Scientist"'s parents as "Profession", and "Profession"'s parents as nothing.
It looks like more of a chore than it is, and it's only for the first time of course.